Disable connection pooling
I use a single client on many tasks, that are most of the time executed
in the same thread, and connection pooling was creating issues
(connections dropped between the moment it is created and when I need to
use it, I think). After many debugging attempts the easiest way I've
found to resolve the issue is to simply create my HTTP client without
connection pooling. For this project, it really does not matter if the
same connection is reused or not when making moves.